BMI category labels using CDC classification

BMI category labels using CDC classification

e_calc_bmi_categories(
  bmi,
  bmi_breaks = c(-Inf, 18.5, 25, 30, Inf),
  bmi_labels = c("Underweight", "Normal or Healthy Weight", "Overweight", "Obese")
)

Arguments

bmi
BMI value

bmi_breaks
Breaks defining labeled ranges, intervals are closed on the left

bmi_labels
Labels for values between break values including lower bound

Value

bmi_cat, categorical levels of BMI

Examples

e_calc_bmi(60, 150) %>% e_calc_bmi_categories()
#> [1] Overweight
#> Levels: Underweight Normal or Healthy Weight Overweight Obese
e_calc_bmi(seq(40, 70, by = 5), 150) %>% e_calc_bmi_categories()
#> [1] Underweight              Normal or Healthy Weight Normal or Healthy Weight
#> [4] Normal or Healthy Weight Overweight               Overweight              
#> [7] Obese                   
#> Levels: Underweight Normal or Healthy Weight Overweight Obese