PHQ-9 category labels using Table 4 from Kroenke K, Spitzer RL, Psychiatric Annals 2002;32:509-521

PHQ-9 Scores and Proposed Treatment Actions. https://www.pcpcc.org/sites/default/files/resources/instructions.pdf

e_calc_phq9_categories(
  phq9,
  phq9_breaks = c(-Inf, 5, 10, 15, 20, Inf),
  phq9_labels = c("None-minimal", "Mild", "Moderate", "Moderately Severe", "Severe")
)

Arguments

phq9
PHQ-9 scores (0 - 27)

phq9_breaks
Breaks defining labeled ranges, intervals are closed on the left

phq9_labels
Labels for values between break values including lower bound

Value

phq9_cat, categorical levels of PHQ-9

Examples

e_calc_phq9_categories(c(0, 6, 12, 21))

#> [1] None-minimal Mild         Moderate     Severe      
#> Levels: None-minimal Mild Moderate Moderately Severe Severe

dplyr::tibble(phq9 = seq(0, 27, by = 1), phq9_cat = phq9 %>% e_calc_phq9_categories()) %>% print(n = Inf)

#> # A tibble: 28 x 2
#>     phq9 phq9_cat         
#>    <dbl> <fct>            
#>  1     0 None-minimal     
#>  2     1 None-minimal     
#>  3     2 None-minimal     
#>  4     3 None-minimal     
#>  5     4 None-minimal     
#>  6     5 Mild             
#>  7     6 Mild             
#>  8     7 Mild             
#>  9     8 Mild             
#> 10     9 Mild             
#> 11    10 Moderate         
#> 12    11 Moderate         
#> 13    12 Moderate         
#> 14    13 Moderate         
#> 15    14 Moderate         
#> 16    15 Moderately Severe
#> 17    16 Moderately Severe
#> 18    17 Moderately Severe
#> 19    18 Moderately Severe
#> 20    19 Moderately Severe
#> 21    20 Severe           
#> 22    21 Severe           
#> 23    22 Severe           
#> 24    23 Severe           
#> 25    24 Severe           
#> 26    25 Severe           
#> 27    26 Severe           
#> 28    27 Severe