Apply labels to a dataset from a table of Var and Label

Apply labels to a dataset from a table of Var and Label

e_data_var_labels_apply_from_table(dat, var_labels)

Arguments

dat

a data.frame

var_labels

a data.frame with columns Var and Label, where Var are column names that may appear in dat

Value

dat the same data.frame with attribute "label" added

Examples

dat <- tibble::tibble(a = 1:10, b = 11:20, c = 21:30)
var_labels <-
  tibble::tribble(
    ~Var, ~Label
  , "a" , "Var a Label"
  , "b" , "Var b Label"
  , "c" , "Var c Label"
  )
dat |> str()
#> tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ a: int [1:10] 1 2 3 4 5 6 7 8 9 10
#>  $ b: int [1:10] 11 12 13 14 15 16 17 18 19 20
#>  $ c: int [1:10] 21 22 23 24 25 26 27 28 29 30
dat |> e_data_var_labels_apply_from_table(var_labels) |> str()
#> tibble [10 × 3] (S3: tbl_df/tbl/data.frame)
#>  $ a: int [1:10] 1 2 3 4 5 6 7 8 9 10
#>   ..- attr(*, "label")= chr "Var a Label"
#>  $ b: int [1:10] 11 12 13 14 15 16 17 18 19 20
#>   ..- attr(*, "label")= chr "Var b Label"
#>  $ c: int [1:10] 21 22 23 24 25 26 27 28 29 30
#>   ..- attr(*, "label")= chr "Var c Label"