Multiple regression power analysis

Multiple regression power analysis

e_lm_power(
  dat = NULL,
  formula_full = NULL,
  formula_red = NULL,
  n_total,
  n_groups,
  sig_level = 0.05,
  weights = NULL,
  sw_print = TRUE,
  sw_plot = TRUE,
  n_plot_ref = NULL
)

Arguments

dat
observed effect size data set

formula_full
observed effect size full model formula

formula_red
observed effect size reduced model formula

n_total
a total sample size value or list of values

n_groups
number of groups for degrees-of-freedom adjustment for Cohen effect sizes

sig_level
Type-I error rate

weights
observed effect size model fit, if it should be weighted regression

sw_print
print results

sw_plot
create plot

n_plot_ref
a sample size reference line for the plot; if null, then uses size of data, otherwise uses median of n_total

Value

Examples

# without data, single n
n_total  <- 100
n_groups <- 3
out <-
  e_lm_power(
    dat           = NULL
  , formula_full  = NULL
  , formula_red   = NULL
  , n_total       = n_total
  , n_groups      = n_groups
  , sig_level     = 0.05
  , weights       = NULL
  , sw_print      = TRUE
  , sw_plot       = TRUE
  , n_plot_ref    = NULL
  )

#> Setting n_plot_ref = n_total for reference

#> Cohen reference effect size and power =============================
#>   Small ---------------
#> 
#>      Multiple regression power calculation 
#> 
#>               u = 2
#>               v = 97
#>              f2 = 0.02
#>       sig.level = 0.05
#>           power = 0.2197557
#> 
#>   Medium --------------
#> 
#>      Multiple regression power calculation 
#> 
#>               u = 2
#>               v = 97
#>              f2 = 0.15
#>       sig.level = 0.05
#>           power = 0.9371108
#> 
#>   Large ---------------
#> 
#>      Multiple regression power calculation 
#> 
#>               u = 2
#>               v = 97
#>              f2 = 0.35
#>       sig.level = 0.05
#>           power = 0.9997727
#> 

#> Error in drop_na(.): could not find function "drop_na"

# without data, sequence of n for power curve
n_total <- seq(10, 300, by = 5)
n_groups <- 3
out <-
  e_lm_power(
    dat           = NULL
  , formula_full  = NULL
  , formula_red   = NULL
  , n_total       = n_total
  , n_groups      = n_groups
  , sig_level     = 0.05
  , weights       = NULL
  , sw_print      = TRUE
  , sw_plot       = TRUE
  , n_plot_ref    = NULL
  )

#> Setting n_plot_ref = median(n_total) for reference

#> Warning: e_lm_power, not printing results when n_total > 1

#> Error in drop_na(.): could not find function "drop_na"

# with data
data(mtcars, package="datasets")
str(mtcars)

#> 'data.frame':	32 obs. of  11 variables:
#>  $ mpg : num  21 21 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 ...
#>  $ cyl : num  6 6 4 6 8 6 8 4 4 6 ...
#>  $ disp: num  160 160 108 258 360 ...
#>  $ hp  : num  110 110 93 110 175 105 245 62 95 123 ...
#>  $ drat: num  3.9 3.9 3.85 3.08 3.15 2.76 3.21 3.69 3.92 3.92 ...
#>  $ wt  : num  2.62 2.88 2.32 3.21 3.44 ...
#>  $ qsec: num  16.5 17 18.6 19.4 17 ...
#>  $ vs  : num  0 0 1 1 0 1 0 1 1 1 ...
#>  $ am  : num  1 1 1 0 0 0 0 0 0 0 ...
#>  $ gear: num  4 4 4 3 3 3 3 4 4 4 ...
#>  $ carb: num  4 4 1 1 2 1 4 2 2 4 ...

yvar      <- "mpg"
xvar_full <- c("cyl", "disp", "hp", "drat", "wt", "qsec")
xvar_red  <- c(               "hp", "drat", "wt", "qsec")

formula_full <-
  as.formula(
    paste0(
      yvar
    , " ~ "
    , paste(
        xvar_full
      , collapse= "+"
      )
    )
  )

formula_red <-
  as.formula(
    paste0(
      yvar
    , " ~ "
    , paste(
        xvar_red
      , collapse= "+"
      )
    )
  )

# with data, single n
n_total  <- 100
n_groups <- 3
out <-
  e_lm_power(
    dat           = datasets::mtcars
  , formula_full  = formula_full
  , formula_red   = formula_red
  , n_total       = n_total
  , n_groups      = n_groups
  , sig_level     = 0.05
  , weights       = NULL
  , sw_print      = TRUE
  , sw_plot       = TRUE
  , n_plot_ref    = NULL
  )

#> Full Model ========================================================
#> mpg ~ cyl + disp + hp + drat + wt + qsec
#> <environment: 0x000000002aa1ea70>
#> 
#> Call:
#> lm(formula = formula_full, data = dat)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.9682 -1.5795 -0.4353  1.1662  5.5272 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)   
#> (Intercept) 26.30736   14.62994   1.798  0.08424 . 
#> cyl         -0.81856    0.81156  -1.009  0.32282   
#> disp         0.01320    0.01204   1.097  0.28307   
#> hp          -0.01793    0.01551  -1.156  0.25846   
#> drat         1.32041    1.47948   0.892  0.38065   
#> wt          -4.19083    1.25791  -3.332  0.00269 **
#> qsec         0.40146    0.51658   0.777  0.44436   
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 2.557 on 25 degrees of freedom
#> Multiple R-squared:  0.8548,	Adjusted R-squared:   0.82 
#> F-statistic: 24.53 on 6 and 25 DF,  p-value: 2.45e-09
#> 
#> Reduced Model =====================================================
#> mpg ~ hp + drat + wt + qsec
#> <environment: 0x000000002aa1ea70>
#> 
#> Call:
#> lm(formula = formula_red, data = dat)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.5775 -1.6626 -0.3417  1.1317  5.4422 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 19.25970   10.31545   1.867 0.072785 .  
#> hp          -0.01784    0.01476  -1.209 0.237319    
#> drat         1.65710    1.21697   1.362 0.184561    
#> wt          -3.70773    0.88227  -4.202 0.000259 ***
#> qsec         0.52754    0.43285   1.219 0.233470    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 2.539 on 27 degrees of freedom
#> Multiple R-squared:  0.8454,	Adjusted R-squared:  0.8225 
#> F-statistic: 36.91 on 4 and 27 DF,  p-value: 1.408e-10
#> 

#> Setting n_plot_ref = the number of observations in the data for reference

#> Warning: e_lm_power, not printing results when n_total > 1

#> Error in drop_na(.): could not find function "drop_na"

# without data, sequence of n for power curve
n_total <- seq(10, 300, by = 5)
n_groups <- 3
out <-
  e_lm_power(
    dat           = datasets::mtcars
  , formula_full  = formula_full
  , formula_red   = formula_red
  , n_total       = n_total
  , n_groups      = n_groups
  , sig_level     = 0.05
  , weights       = NULL
  , sw_print      = TRUE
  , sw_plot       = TRUE
  , n_plot_ref    = 100
  )

#> Full Model ========================================================
#> mpg ~ cyl + disp + hp + drat + wt + qsec
#> <environment: 0x000000002aa1ea70>
#> 
#> Call:
#> lm(formula = formula_full, data = dat)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.9682 -1.5795 -0.4353  1.1662  5.5272 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)   
#> (Intercept) 26.30736   14.62994   1.798  0.08424 . 
#> cyl         -0.81856    0.81156  -1.009  0.32282   
#> disp         0.01320    0.01204   1.097  0.28307   
#> hp          -0.01793    0.01551  -1.156  0.25846   
#> drat         1.32041    1.47948   0.892  0.38065   
#> wt          -4.19083    1.25791  -3.332  0.00269 **
#> qsec         0.40146    0.51658   0.777  0.44436   
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 2.557 on 25 degrees of freedom
#> Multiple R-squared:  0.8548,	Adjusted R-squared:   0.82 
#> F-statistic: 24.53 on 6 and 25 DF,  p-value: 2.45e-09
#> 
#> Reduced Model =====================================================
#> mpg ~ hp + drat + wt + qsec
#> <environment: 0x000000002aa1ea70>
#> 
#> Call:
#> lm(formula = formula_red, data = dat)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.5775 -1.6626 -0.3417  1.1317  5.4422 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept) 19.25970   10.31545   1.867 0.072785 .  
#> hp          -0.01784    0.01476  -1.209 0.237319    
#> drat         1.65710    1.21697   1.362 0.184561    
#> wt          -3.70773    0.88227  -4.202 0.000259 ***
#> qsec         0.52754    0.43285   1.219 0.233470    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> 
#> Residual standard error: 2.539 on 27 degrees of freedom
#> Multiple R-squared:  0.8454,	Adjusted R-squared:  0.8225 
#> F-statistic: 36.91 on 4 and 27 DF,  p-value: 1.408e-10
#> 

#> Warning: e_lm_power, not printing results when n_total > 1

#> Error in drop_na(.): could not find function "drop_na"

### RMarkdown results reporting
# The results above indicate the following.
#
# 1. With $n=`r out[["tab_power"]] %>% filter(n_total == 100) %>% pull(n_total)`$
#    and number of groups $k=`r out[["tab_power"]] %>% filter(n_total == 100) %>% pull(n_groups)`$
# 2. prelim power = `r out[["tab_power"]] %>% filter(n_total == 100) %>% pull(obs_power) %>% signif(digits = 2)`
# 3. Cohen small, medium, and large power:
#     * `r out[["tab_power"]] %>% filter(n_total == 100) %>% pull(Cohen_small_power ) %>% signif(digits = 2)`,
#     * `r out[["tab_power"]] %>% filter(n_total == 100) %>% pull(Cohen_medium_power) %>% signif(digits = 2)`, and
#     * `r out[["tab_power"]] %>% filter(n_total == 100) %>% pull(Cohen_large_power ) %>% signif(digits = 2)`.