Split a (set of) item-listing columns into indicator columns

Commonly used for comorbidities or prescription lists within a single (or multiple) column(s). Takes a column where items are separated by puncutation (,.;/|) and creates separate columns with indicators. Can treat counts of items >1 as 1 to simplify summary tables.

e_split_list_columns_into_indicator_columns(
  dat_this,
  var_names_items = NULL,
  code_other_below_freq = 5,
  label_other = "other",
  indicator_col_prefix = "item_",
  sw_replace_GT1_with_1 = FALSE,
  sw_print_unique = TRUE
)

Arguments

dat_this
entire data.frame or tibble, will return with additional indicator columns

var_names_items
column names with lists of items

code_other_below_freq
replace item name with label_other if total frequency for an item is less than this value

label_other
label for the "other" category

indicator_col_prefix
prefix for the new indicator columns

sw_replace_GT1_with_1
if a person's item count is more than one in the indicator column, should we replace with a "1" to interpret as "at least 1"

sw_print_unique
print list of items before and after replacing with "other"

Value

dat_this with additional indicator columns

Examples

dat_ex <-
  dplyr::tibble(
    col1 = c(NA, "", "a", "A, B  ,C", "b", "D. c ;    d", "x  ;Y", "ab/0|J", 2, "other", 1)
  , col2 = LETTERS[1:length(col1)]
  ) %>%
  dplyr::mutate(
    ID = 1:dplyr::n()
  ) %>%
  dplyr::select(
    ID
  , everything()
  )
dat_ex %>% print(n = Inf)

#> # A tibble: 11 x 3
#>       ID col1          col2 
#>    <int> <chr>         <chr>
#>  1     1  NA           A    
#>  2     2 ""            B    
#>  3     3 "a"           C    
#>  4     4 "A, B  ,C"    D    
#>  5     5 "b"           E    
#>  6     6 "D. c ;    d" F    
#>  7     7 "x  ;Y"       G    
#>  8     8 "ab/0|J"      H    
#>  9     9 "2"           I    
#> 10    10 "other"       J    
#> 11    11 "1"           K    

dat_ex_out <-
  e_split_list_columns_into_indicator_columns(
    dat_this              = dat_ex
  , var_names_items       = c("col1", "col2")
  , code_other_below_freq = 2
  , label_other           = "other"
  , indicator_col_prefix  = "item_"
  , sw_replace_GT1_with_1 = FALSE
  , sw_print_unique       = TRUE
  )

#> [1] "Unique items: 18"
#> # A tibble: 18 x 2
#>    items  Freq
#>    <chr> <int>
#>  1 a         3
#>  2 b         3
#>  3 c         3
#>  4 d         3
#>  5 j         2
#>  6 0         1
#>  7 1         1
#>  8 2         1
#>  9 ab        1
#> 10 e         1
#> 11 f         1
#> 12 g         1
#> 13 h         1
#> 14 i         1
#> 15 k         1
#> 16 other     1
#> 17 x         1
#> 18 y         1
#> [1] "Coding items with frequencies less than 2 to 'other'"

#> erikmisc::e_split_list_columns_into_indicator_columns: Other category `other` already appears in data.

#> [1] "Unique items: 6 with other category `other`"
#> # A tibble: 6 x 2
#>   items  Freq
#>   <chr> <int>
#> 1 other    13
#> 2 a         3
#> 3 b         3
#> 4 c         3
#> 5 d         3
#> 6 j         2

#> New names:
#> * NA -> ...1
#> * NA -> ...2

dat_ex_out %>% print(n = Inf)

#> # A tibble: 11 x 9
#>       ID col1          col2  item_other item_a item_b item_c item_d item_j
#>    <int> <chr>         <chr>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1     1  NA           A              0      1      0      0      0      0
#>  2     2 ""            B              0      0      1      0      0      0
#>  3     3 "a"           C              0      1      0      1      0      0
#>  4     4 "A, B  ,C"    D              0      1      1      1      1      0
#>  5     5 "b"           E              1      0      1      0      0      0
#>  6     6 "D. c ;    d" F              1      0      0      1      2      0
#>  7     7 "x  ;Y"       G              3      0      0      0      0      0
#>  8     8 "ab/0|J"      H              3      0      0      0      0      1
#>  9     9 "2"           I              2      0      0      0      0      0
#> 10    10 "other"       J              1      0      0      0      0      1
#> 11    11 "1"           K              2      0      0      0      0      0