Print a data.frame table for latex and scale to fit page

If using in Rmd file and a latex documnet, then requires two header includes:

e_table_print(
  dat,
  sw_scale = FALSE,
  sw_kable_format = c("simple", "kbl", "html", "latex", "doc")[2],
  sw_latex_options = c("basic", "striped", "hold_position", "HOLD_position",
    "scale_down", "repeat_header")[c(2, 3)],
  ...
)

Arguments

dat

data.frame or tibble to print

sw_scale

if "latex" T/F to scale to fit page, if "html" then a point size to pass to kableExtra::kable_styling argument font_size

sw_kable_format

"simple", "html", "latex", or "doc"

sw_latex_options

passed to kableExtra::kable_styling(latex_options = sw_latex_options), updated if sw_scale=TRUE to include "scale_down"

...

other arguments passed to knitr::kable()

Value

invisible(NULL)

Details

EXAMPLE OF .Rmd YAML header for pdf via LaTeX

---
title: "Title"
date: "Today"
output:
  pdf_document:
    df_print: kable
    latex_engine: lualatex
    toc: true
    number_sections: true
    toc_depth: 2
    keep_tex: true
header-includes:
- \usepackage{booktabs}
- \usepackage{colortbl}
always_allow_html: yes
fig_caption: TRUE
---

Examples

dat_mtcars_e |> head() |> e_table_print()
#> <table class=" lightable-paper lightable-hover" style='font-family: "Arial Narrow", arial, helvetica, sans-serif; width: auto !important; margin-left: auto; margin-right: auto;'>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> model </th>
#>    <th style="text-align:right;"> mpg </th>
#>    <th style="text-align:left;"> cyl </th>
#>    <th style="text-align:right;"> disp </th>
#>    <th style="text-align:right;"> hp </th>
#>    <th style="text-align:right;"> drat </th>
#>    <th style="text-align:right;"> wt </th>
#>    <th style="text-align:right;"> qsec </th>
#>    <th style="text-align:left;"> vs </th>
#>    <th style="text-align:left;"> am </th>
#>    <th style="text-align:right;"> gear </th>
#>    <th style="text-align:right;"> carb </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;"> Mazda RX4 </td>
#>    <td style="text-align:right;"> 21.0 </td>
#>    <td style="text-align:left;"> six </td>
#>    <td style="text-align:right;"> 160 </td>
#>    <td style="text-align:right;"> 110 </td>
#>    <td style="text-align:right;"> 3.90 </td>
#>    <td style="text-align:right;"> 2.620 </td>
#>    <td style="text-align:right;"> 16.46 </td>
#>    <td style="text-align:left;"> V-shaped </td>
#>    <td style="text-align:left;"> manual </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 4 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Mazda RX4 Wag </td>
#>    <td style="text-align:right;"> 21.0 </td>
#>    <td style="text-align:left;"> six </td>
#>    <td style="text-align:right;"> 160 </td>
#>    <td style="text-align:right;"> 110 </td>
#>    <td style="text-align:right;"> 3.90 </td>
#>    <td style="text-align:right;"> 2.875 </td>
#>    <td style="text-align:right;"> 17.02 </td>
#>    <td style="text-align:left;"> V-shaped </td>
#>    <td style="text-align:left;"> manual </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 4 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Datsun 710 </td>
#>    <td style="text-align:right;"> 22.8 </td>
#>    <td style="text-align:left;"> four </td>
#>    <td style="text-align:right;"> 108 </td>
#>    <td style="text-align:right;"> 93 </td>
#>    <td style="text-align:right;"> 3.85 </td>
#>    <td style="text-align:right;"> 2.320 </td>
#>    <td style="text-align:right;"> 18.61 </td>
#>    <td style="text-align:left;"> straight </td>
#>    <td style="text-align:left;"> manual </td>
#>    <td style="text-align:right;"> 4 </td>
#>    <td style="text-align:right;"> 1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Hornet 4 Drive </td>
#>    <td style="text-align:right;"> 21.4 </td>
#>    <td style="text-align:left;"> six </td>
#>    <td style="text-align:right;"> 258 </td>
#>    <td style="text-align:right;"> 110 </td>
#>    <td style="text-align:right;"> 3.08 </td>
#>    <td style="text-align:right;"> 3.215 </td>
#>    <td style="text-align:right;"> 19.44 </td>
#>    <td style="text-align:left;"> straight </td>
#>    <td style="text-align:left;"> automatic </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 1 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Hornet Sportabout </td>
#>    <td style="text-align:right;"> 18.7 </td>
#>    <td style="text-align:left;"> eight </td>
#>    <td style="text-align:right;"> 360 </td>
#>    <td style="text-align:right;"> 175 </td>
#>    <td style="text-align:right;"> 3.15 </td>
#>    <td style="text-align:right;"> 3.440 </td>
#>    <td style="text-align:right;"> 17.02 </td>
#>    <td style="text-align:left;"> V-shaped </td>
#>    <td style="text-align:left;"> automatic </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 2 </td>
#>   </tr>
#>   <tr>
#>    <td style="text-align:left;"> Valiant </td>
#>    <td style="text-align:right;"> 18.1 </td>
#>    <td style="text-align:left;"> six </td>
#>    <td style="text-align:right;"> 225 </td>
#>    <td style="text-align:right;"> 105 </td>
#>    <td style="text-align:right;"> 2.76 </td>
#>    <td style="text-align:right;"> 3.460 </td>
#>    <td style="text-align:right;"> 20.22 </td>
#>    <td style="text-align:left;"> straight </td>
#>    <td style="text-align:left;"> automatic </td>
#>    <td style="text-align:right;"> 3 </td>
#>    <td style="text-align:right;"> 1 </td>
#>   </tr>
#> </tbody>
#> </table>
if (FALSE) {
# html rescaled size
dat_mtcars_e |>
  head() |>
  e_table_print(sw_scale = 6)
# outputs into LaTeX document, scaling is automatic to fit page width
# works best before a \clearpage
# use chunk option: results = 'asis', see other options at https://yihui.org/knitr/options/
dat_mtcars_e |>
  head() |>
  e_table_print(sw_scale = TRUE, sw_kable_format = "latex")
}