Random Forests classification workflow using randomForestSRC

Random Forests classification workflow using randomForestSRC

e_rfsrc_classification(
  dat_rf_class = NULL,
  rf_y_var = NULL,
  rf_x_var = NULL,
  rf_id_var = NULL,
  sw_rfsrc_ntree = 2000,
  sw_alpha = 0.05,
  sw_save_model = c(TRUE, FALSE)[1],
  plot_title = "Random Forest Title",
  out_path = "out",
  file_prefix = "out_e_rf",
  plot_format = c("png", "pdf")[1]
)

Arguments

dat_rf_class

data.frame with data

rf_y_var

y factor variable, if NULL then assumed to be the first column of dat_rf_class

rf_x_var

x variables, if NULL then assumed to be all except the first column of dat_rf_class

rf_id_var

ID variable, removed from dataset prior to analysis

sw_rfsrc_ntree

ntree argument for randomForestSRC::rfsrc()

sw_alpha

alpha argument for randomForestSRC::plot.subsample() and randomForestSRC::extract.bootsample()

sw_save_model

T/F to save model (XXX, not yet implemented)

plot_title

title for plots

out_path

path to save output

file_prefix

file prefix for saved output

plot_format

plot format supported by ggplot2::ggsave()

Value

list with many RF objects, summaries, and plots

Examples

dat_rf_class <-
  erikmisc::dat_mtcars_e |>
  dplyr::select(
    -model
  ) |>
  #dplyr::select(
  #  cyl
  #, tidyselect::everything()
  #) |>
  dplyr::mutate(
    ID    = 1:dplyr::n() # ID number
  , disp  =              # add missing values
      dplyr::case_when(
        disp |> dplyr::between(160, 170) ~ NA
      , TRUE ~ disp
      )
  )

out_e_rf <-
  e_rfsrc_classification(
    dat_rf_class    = dat_rf_class
  , rf_y_var        = "cyl"    # NULL
  , rf_x_var        = c("mpg", "disp", "hp", "drat", "wt", "qsec", "vs", "am") # NULL
  , rf_id_var       = "ID"
  , sw_rfsrc_ntree  = 2000
  , sw_alpha        = 0.05
  , sw_save_model   = c(TRUE, FALSE)[1]
  , plot_title      = "Random Forest Title"
  , out_path        = "."
  , file_prefix     = "out_e_rf"
  , plot_format     = c("png", "pdf")[1]
  )
#> [1] "erikmisc::e_rfsrc_classification, 0s:  Create output directory"
#> [1] "erikmisc::e_rfsrc_classification, 0.08s:  Removing rows with NAs (later versions will allow imputations)"
#> [1] "erikmisc::e_rfsrc_classification, 0.1s:  Grow forest"
#> 
#> Trees Grown:       1,    Time Remaining (sec):       0 
#> 
#> 
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> [1] "erikmisc::e_rfsrc_classification, 5.48s:  Subsample for VIMP and model selection"
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%
#> Warning: empty classes found when implementing classification
#>   |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%
#> Warning: empty classes found when implementing classification
#>   |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%
#> Warning: empty classes found when implementing classification
#>   |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> [1] "erikmisc::e_rfsrc_classification, 45.59s:  Model selection"
#> [1] "erikmisc::e_rfsrc_classification, 45.64s:  Grow forest (selected model)"
#> 
#> Trees Grown:       1,    Time Remaining (sec):       0 
#> 
#> 
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> [1] "erikmisc::e_rfsrc_classification, 51.07s:  Subsample for VIMP with CIs and SEs (selected model)"
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%
#> Warning: empty classes found when implementing classification
#>   |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%
#> Warning: empty classes found when implementing classification
#>   |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%
#> Warning: empty classes found when implementing classification
#>   |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> [1] "erikmisc::e_rfsrc_classification, 87.47s (~1.46 minutes):  Partial effects plots (selected model)"
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> Warning: Package `gridGraphics` is required to handle base-R plots. Substituting empty plot.
#> [1] "erikmisc::e_rfsrc_classification, 90.72s (~1.51 minutes):  ROC Curve (selected model)"
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Coordinate system already present. Adding new coordinate system, which will
#> replace the existing one.
#> Error in out[["plot_o_class"]] + labs(title = NULL): non-numeric argument to binary operator

if (FALSE) {

## Overall summaries

# Summary of Full and reduced model fits with ROC curves
out_e_rf[[ "plot_rf_train_all_summary" ]]

# Reduced model: ROC object from e_plot_roc()
out_e_rf[[ "plot_o_class_sel_ROC" ]]$plot_roc |>
  cowplot::plot_grid(plotlist = _, nrow = 1)

# Reduced model: Marginal/Partial effects plots
  # for each level of the response variable
for (i_level in seq_along(levels(dat_rf_class[[ out_e_rf[[ "rf_y_var" ]] ]]))) {
  out_e_rf[[ "plot_o_class_sel_marginal_effects" ]][[ i_level ]] |>
    cowplot::plot_grid() |>
    print()
}

## Full model summaries
# y response variable
out_e_rf[[ "rf_y_var" ]]
# Full model: x predictor variables
out_e_rf[[ "rf_x_var" ]]
# Full model: formula
out_e_rf[[ "rf_formula" ]]
# Full model: rfsrc classification object
out_e_rf[[ "o_class" ]]
# Full model: convergence
out_e_rf[[ "plot_o_class" ]] |>
  cowplot::plot_grid()
# Full model: variable importance (VIMP) table
out_e_rf[[ "o_class_importance" ]]
# Full model: variable importance (VIMP) plot
out_e_rf[[ "plot_o_class_importance" ]]
# Full model: ROC AUC (area under the curve)
out_e_rf[[ "o_class_AUC" ]]
# Full model: subsample iterates for VIMP and model selection
out_e_rf[[ "o_class_subsample" ]]
# Full model: subsample VIMP model selection
out_e_rf[[ "o_class_subsample_extract_subsample" ]]
# Full model: subsample double bootstrap VIMP model selection
out_e_rf[[ "o_class_subsample_extract_bootsample" ]]
# Full model: variable importance (VIMP) plot boxplots
out_e_rf[[ "plot_o_class_subsample" ]] |>
  cowplot::plot_grid()

## Reduced model summaries
# Reduced model: x predictor variables
out_e_rf[[ "rf_x_var_sel" ]]
# Reduced model: formula
out_e_rf[[ "rf_formula_sel" ]]
# Reduced model: rfsrc classification object
out_e_rf[[ "o_class_sel" ]]
# Reduced model: convergence
out_e_rf[[ "plot_o_class_sel" ]] |>
  cowplot::plot_grid()
# Reduced model: variable importance (VIMP) table
out_e_rf[[ "o_class_sel_importance" ]]
# Reduced model: variable importance (VIMP) plot
out_e_rf[[ "plot_o_class_sel_importance" ]]
# Reduced model: ROC AUC (area under the curve)
out_e_rf[[ "o_class_sel_AUC" ]]
# Reduced model: subsample iterates for VIMP and model selection
out_e_rf[[ "o_class_sel_subsample" ]]
# Reduced model: subsample VIMP model selection
out_e_rf[[ "o_class_sel_subsample_extract_subsample" ]]
# Reduced model: subsample double bootstrap VIMP model selection
out_e_rf[[ "o_class_sel_subsample_extract_bootsample" ]]
# Reduced model: variable importance (VIMP) plot boxplots
out_e_rf[[ "plot_o_class_sel_subsample" ]] |>
  cowplot::plot_grid()

}