Replace Unicode with ASCII

Read a text file and replace unicode characters [^\x00-\x7F] with equivalent ASCII or specified replacement. Review table ?dat_unicode_tables_sel_byhand_e for the ASCII replacements. Not all unicode has a natural ASCII replacement, so those replacements can be specified with unicode_generic_replacement.

e_text_unicode_to_ascii(
  fn_in = NULL,
  fn_out = NULL,
  unicode_generic_replacement = "XunicodeX"
)

Arguments

fn_in
text filename in.

fn_out
text filename out. If NULL, then this is assigned fn_in.

unicode_generic_replacement
a string to replace non-matched unicode characters

Value

NULL, invisibly

Examples

if (FALSE) {
# e_text_unicode_to_ascii(
#     fn_in  = "./data-raw/text_in_unicode.csv"
#   , fn_out = "./data-raw/text_out_unicode.csv"
#   , unicode_generic_replacement = "XxXunicodeXxX"
#   )
}